requirements.txt
pandas
pydantic
firecrawl-py
selenium
webdriver-manager


prefect